{{cookiecutter.github_repository_name}}/requirements/base.txt
# Django
Django==1.9.7

# Environmental var configuration
django-configurations==1.0

# For the persistence stores
psycopg2==2.6.1
dj-database-url==0.4.0

# Models
django-model-utils==2.4

# Rest apis
djangorestframework==3.3.3
Markdown==2.6.6
django-filter==0.13.0

# Static files
whitenoise==3.0

# Media files
boto==2.39.0
django_unique_upload==0.2.1

# Image manipulation
django-versatileimagefield==1.3

# For asynchronous queuing
django-rq==0.9.0

# Time zones support
pytz==2016.3

# Admion Theme
django-suit==0.2.18

